Split start-account-or-login/steps.tsx into many files.
